refactored Profile section
